# docs/architecture/reflexion-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Reflexion Platform API
  version: 1.0.0
  description: >-
    Implemented unified contract for Mirror, Caregiver, Provider, device pairing,
    session ingestion, care plans, longitudinal monitoring, and review workflows.
    Long-running media and ML processing is asynchronous.
servers:
  - url: https://api.example.com
    description: Self-hosted production API origin
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/patients/{patientId}/care-plan:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [CarePlan]
      summary: Get Care Plan
      operationId: getCarePlan
      security:
        - bearerAuth: []
        - deviceBearer: []
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        default: { $ref: "#/components/responses/Error" }
    put:
      tags: [CarePlan]
      summary: Replace Care Plan
      operationId: replaceCarePlan
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/patients/{patientId}/medication-plans:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [CarePlan]
      summary: List Medication Plans
      operationId: listMedicationPlans
      security:
        - bearerAuth: []
        - deviceBearer: []
      responses:
        "200":
          description: Active medication plans visible to the actor.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MedicationPlan"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [CarePlan]
      summary: Create Medication Plan
      operationId: createMedicationPlan
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationPlanRequest"
      responses:
        "201":
          description: Medication plan created by an authorized human actor.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/MedicationPlan"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/patients/{patientId}/reminder-occurrences:
    get:
      tags: [CarePlan]
      summary: List Reminder Occurrences
      operationId: listReminderOccurrences
      security:
        - bearerAuth: []
        - deviceBearer: []
      parameters:
        - $ref: "#/components/parameters/PatientId"
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Reminder occurrences in the requested time window.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReminderOccurrence"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/medication-plans/{planId}:
    patch:
      tags: [CarePlan]
      summary: Update Medication Plan
      operationId: updateMedicationPlan
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/patients/{patientId}/caregiver-tasks:
    post:
      tags: [CarePlan]
      summary: Create Caregiver Task
      operationId: createCaregiverTask
      security:
        - bearerAuth: []
        - deviceBearer: []
      parameters:
        - $ref: "#/components/parameters/PatientId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, priority, title]
              properties:
                category: { type: string, enum: [follow_up, appointment, medication_review, technical, custom] }
                priority: { type: string, enum: [routine, elevated, urgent] }
                title: { type: string, maxLength: 160 }
                details: { type: string, maxLength: 1000 }
                dueAt: { type: string, format: date-time }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/reminder-occurrences/{occurrenceId}/responses:
    post:
      tags: [CarePlan]
      summary: Record Reminder Response
      operationId: recordReminderResponse
      security:
        - bearerAuth: []
        - deviceBearer: []
      parameters:
        - name: occurrenceId
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, respondedAt]
              properties:
                status:
                  type: string
                  enum: [taken, skipped, snoozed, unknown]
                respondedAt:
                  type: string
                  format: date-time
                note:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Response recorded idempotently.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/ReminderOccurrence"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Caregiver/provider access token.
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Device-scoped short-lived access token.
    deviceBootstrap:
      type: apiKey
      in: header
      name: X-Device-Bootstrap
      description: Installation credential used only before device pairing.
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
  responses:
    DataResponse:
      description: Standard Reflexion data envelope.
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data:
                description: Route-specific response data.
              meta:
                $ref: "#/components/schemas/ResponseMeta"
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: array
              items:
                type: object
                additionalProperties: true
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    MedicationPlan:
      type: object
      required: [planId, patientId, displayName, schedule, status, version]
      properties:
        planId:
          type: string
        patientId:
          type: string
        displayName:
          type: string
        instructions:
          type: string
        schedule:
          type: object
          required: [timezone, times]
          properties:
            timezone:
              type: string
            times:
              type: array
              items:
                type: string
                pattern: "^[0-2][0-9]:[0-5][0-9]$"
            recurrence:
              type: string
        source:
          type: string
          enum: [caregiver, provider]
        status:
          type: string
          enum: [active, paused, ended]
        version:
          type: integer
    CreateMedicationPlanRequest:
      type: object
      required: [displayName, schedule, source]
      properties:
        displayName:
          type: string
          maxLength: 160
        instructions:
          type: string
          maxLength: 1000
        schedule:
          type: object
          required: [timezone, times]
          properties:
            timezone:
              type: string
            times:
              type: array
              minItems: 1
              items:
                type: string
                pattern: "^[0-2][0-9]:[0-5][0-9]$"
            recurrence:
              type: string
        source:
          type: string
          enum: [caregiver, provider]
    ReminderOccurrence:
      type: object
      required: [occurrenceId, patientId, scheduledAt, type, status]
      properties:
        occurrenceId:
          type: string
        patientId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [medication, appointment, activity, custom]
        displayText:
          type: string
        status:
          type: string
          enum: [scheduled, delivered, taken, skipped, snoozed, missed, cancelled, unknown]
        respondedAt:
          type: [string, "null"]
          format: date-time
...